fix: add onAction + minor changes in directories
