+++ src/main/database/DoctorDatabase.csv
@@ -1,3 +1,3 @@
 username,password,name,specialty,stars
-Ryujin,123,Ryujin,Cardiology,5
+Ryujin,123,Ryujin,Cardiology,3
 nayeon,123,nayeon,Cardiology,3
