+++ src/main/java/com/example/clinic/homeSystem/patientHome/PatientEditController.java
@@ -1,0 +1,5 @@
+package com.example.clinic.homeSystem.patientHome;
+
+public class PatientEditController {
+
+}
