+++ src/main/resources/com/example/clinic/DoctorHomeScene/doctorhome-view.fxml
@@ -5,7 +5,7 @@
 <?import javafx.scene.shape.*?>
 <?import javafx.scene.text.*?>
 
-<BorderPane xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.clinic.homeSystem.DoctorHomeController" stylesheets="@styles.css" prefWidth="1080.0" prefHeight="720.0">
+<BorderPane xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.clinic.homeSystem.doctorHome.DoctorHomeController" stylesheets="@styles.css" prefWidth="1080.0" prefHeight="720.0">
    <left>
       <VBox styleClass="sidebar" prefWidth="300.0" alignment="TOP_CENTER">
          <VBox styleClass="profile-cluster" alignment="CENTER" spacing="15.0">
