+++ src/main/resources/com/example/clinic/PatientHomeScene/cancelappointment/PatientCancelAppointments.fxml
