+++ src/main/resources/com/example/clinic/PatientHomeScene/editinfo/editinfo.fxml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import java.lang.*?>
+<?import java.util.*?>
+<?import javafx.scene.*?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+
+<AnchorPane xmlns="http://javafx.com/javafx"
+            xmlns:fx="http://javafx.com/fxml"
+            fx:controller="com.example.clinic.homeSystem.patientHome.PatientEditController"
+            prefHeight="400.0" prefWidth="600.0">
+
+</AnchorPane>
