+++ src/main/resources/com/example/clinic/PatientHomeScene/home/patienthome.fxml
@@ -1,0 +1,121 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.Region?>
+<?import javafx.scene.layout.StackPane?>
+<?import javafx.scene.layout.VBox?>
+<?import javafx.scene.shape.Circle?>
+<?import javafx.scene.shape.Rectangle?>
+<?import javafx.scene.text.Font?>
+
+<HBox prefHeight="720.0" prefWidth="1080.0" stylesheets="@styles.css" xmlns="http://javafx.com/javafx/23.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.example.clinic.homeSystem.PatientHomeController">
+        <children>
+                <!-- Sidebar Verde -->
+                <VBox alignment="CENTER" prefWidth="450.0" spacing="20.0" styleClass="sidebar">
+                        <padding>
+                                <Insets bottom="60.0" left="40.0" right="40.0" top="60.0" />
+                        </padding>
+                        <children>
+                                <StackPane prefHeight="120" prefWidth="120">
+                                        <children>
+                                                <Circle radius="60.0" styleClass="profile-image-background" />
+                                                <VBox alignment="CENTER" spacing="3" style="-fx-translate-y: 5;">
+                                                        <children>
+                                                                <Circle radius="18" styleClass="profile-icon" />
+                                                                <Circle radius="30" styleClass="profile-icon">
+                                                                        <clip>
+                                                                                <Rectangle height="30" width="60" x="-30" y="0" />
+                                                                        </clip>
+                                                                </Circle>
+                                                        </children>
+                                                </VBox>
+                                        </children>
+                                </StackPane>
+
+                                <!-- Informações do Paciente -->
+                                <VBox alignment="CENTER" spacing="8.0" styleClass="patient-info">
+                                        <VBox.margin>
+                                                <Insets bottom="20.0" />
+                                        </VBox.margin>
+                                        <children>
+                                                <Label fx:id="patientNameLabel" styleClass="patient-name" text="Nome do Paciente">
+                                                        <font>
+                                                                <Font name="System Bold" size="20.0" />
+                                                        </font>
+                                                </Label>
+                                                <Label fx:id="patientPlanLabel" styleClass="patient-plan" text="Insurance">
+                                                        <font>
+                                                                <Font size="16.0" />
+                                                        </font>
+                                                </Label>
+                                        </children>
+                                </VBox>
+
+                                <!-- Botões do Menu -->
+                                <VBox maxWidth="320.0" spacing="20.0" styleClass="menu-buttons">
+                                        <VBox.margin>
+                                                <Insets bottom="20.0" />
+                                        </VBox.margin>
+                                        <children>
+                                                <Button fx:id="scheduleBtn" onAction="#handleScheduleAppointment" maxWidth="Infinity" styleClass="menu-btn" text="Schedule Appointment">
+                                                        <font>
+                                                                <Font name="System Bold" size="14.0" />
+                                                        </font>
+                                                </Button>
+                                                <Button fx:id="cancelBtn" onAction="#handleCancelAppointment" maxWidth="Infinity" styleClass="menu-btn" text="Cancel Appointment">
+                                                        <font>
+                                                                <Font name="System Bold" size="14.0" />
+                                                        </font>
+                                                </Button>
+                                                <Button fx:id="evaluateBtn" onAction="#handleReviewDoctor" maxWidth="Infinity" styleClass="menu-btn" text="Evaluete Doctor">
+                                                        <font>
+                                                                <Font name="System Bold" size="14.0" />
+                                                        </font>
+                                                </Button>
+                                        </children>
+                                </VBox>
+
+                                <!-- Spacer para empurrar os botões de perfil para baixo -->
+                                <Region VBox.vgrow="ALWAYS" />
+
+                                <!-- Seção de Perfil -->
+                                <VBox maxWidth="320.0" spacing="15.0" styleClass="profile-section">
+                                        <children>
+                                                <Button fx:id="editProfileBtn" maxWidth="Infinity" styleClass="edit-profile-btn" text="Edit Profile">
+                                                        <font>
+                                                                <Font size="14.0" />
+                                                        </font>
+                                                </Button>
+                                                <Button fx:id="logoutBtn" onAction="#handleLogout" maxWidth="Infinity" styleClass="logout-btn" text="Logout">
+                                                        <font>
+                                                                <Font size="14.0" />
+                                                        </font>
+                                                </Button>
+                                        </children>
+                                </VBox>
+                        </children>
+                </VBox>
+
+                <!-- Área Principal Branca -->
+                <VBox alignment="CENTER" spacing="20.0" styleClass="main-content" HBox.hgrow="ALWAYS">
+                        <padding>
+                                <Insets bottom="60.0" left="80.0" right="80.0" top="60.0" />
+                        </padding>
+                        <children>
+                                <Label styleClass="welcome-title" text="Welcome to the">
+                                        <font>
+                                                <Font name="System Bold" size="48.0" />
+                                        </font>
+                                </Label>
+                                <Label styleClass="welcome-subtitle" text="Patient Home">
+                                        <font>
+                                                <Font name="System Bold" size="48.0" />
+                                        </font>
+                                </Label>
+                        </children>
+                </VBox>
+        </children>
+</HBox>
