--- src/main/resources/com/example/clinic/PatientHomeScene/patienthome.fxml
@@ -1,62 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.scene.shape.*?>
-<?import javafx.scene.text.*?>
-
-<BorderPane xmlns="http://javafx.com/javafx/11.0.1"
-            xmlns:fx="http://javafx.com/fxml/1"
-            fx:controller="com.example.clinic.homeSystem.PatientHomeController"
-            stylesheets="@styles.css"
-            prefWidth="1080.0" prefHeight="720.0">
-
-        <left>
-                <VBox styleClass="sidebar" prefWidth="300.0" alignment="TOP_CENTER">
-                        <VBox styleClass="profile-cluster" alignment="CENTER" spacing="15.0">
-                                <StackPane prefHeight="120" prefWidth="120">
-                                        <children>
-                                                <Circle styleClass="profile-image-background" radius="60.0" />
-                                                <VBox alignment="CENTER" spacing="3" style="-fx-translate-y: 5;">
-                                                        <children>
-                                                                <Circle styleClass="profile-icon" radius="18" />
-                                                                <Circle styleClass="profile-icon" radius="30">
-                                                                        <clip>
-                                                                                <Rectangle x="-30" y="0" width="60" height="30" />
-                                                                        </clip>
-                                                                </Circle>
-                                                        </children>
-                                                </VBox>
-                                        </children>
-                                </StackPane>
-
-                                <Label fx:id="patientNameLabel" styleClass="doctor-name" text="Patient Name" />
-
-
-                        </VBox>
-
-                        <Region VBox.vgrow="ALWAYS" />
-
-                        <Button fx:id="logoutButton" onAction="#handleLogout" styleClass="logout-btn" text="Logout" />
-                </VBox>
-        </left>
-
-        <center>
-                <VBox styleClass="main-content" spacing="30.0">
-                        <VBox styleClass="header-section" spacing="10.0">
-                                <Label fx:id="welcomeLabel" styleClass="welcome-title" text="Welcome, [Patient Name]!" />
-                                <Label fx:id="dateLabel" styleClass="date-label" text="Today is: [Date]" />
-                        </VBox>
-
-                        <VBox spacing="20" styleClass="appointments-section">
-                                <Label styleClass="section-title" text="What would you like to do?" />
-
-                                <VBox spacing="15">
-                                        <Button fx:id="scheduleButton" onAction="#handleScheduleAppointment" text="Schedule Appointment" styleClass="all-appointments-btn" />
-                                        <Button fx:id="cancelButton" onAction="#handleCancelAppointment" text="Cancel Appointment" styleClass="all-appointments-btn" />
-                                        <Button fx:id="reviewButton" onAction="#handleReviewDoctor" text="Evaluate Doctor" styleClass="all-appointments-btn" />
-                                </VBox>
-                        </VBox>
-                </VBox>
-        </center>
-</BorderPane>
